APM-6127 remove old fixtures
